pyproject.toml
[tool.poetry]
name = "rag-eval"
version = "0.1.0"
description = "Production-grade RAG system with rigorous evaluation framework"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "rag_eval", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
llama-index-core = "^0.10.0"
llama-index-vector-stores-qdrant = "^0.2.0"
sentence-transformers = "^2.2.0"
qdrant-client = "^1.7.0"
rank-bm25 = "^0.2.2"
beir = "^2.0.0"
openai = "^1.0.0"
anthropic = "^0.18.0"
httpx = "^0.25.0"
ragas = "^0.1.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
numpy = "^1.24.0"
pandas = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
ruff = "^0.1.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
